requirements.txt
aiodns==3.1.1
aiosignal==1.3.1
alpaca-py==0.15.0
alpha-vantage==2.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.10.4
astrapy==0.7.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.22.1
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
cassandra-driver==3.29.0
cassio==0.1.4
ccxt==4.1.78
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.2
cssselect==1.2.0
curl_cffi==0.6.0b9
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.5.0
docstring-inheritance==2.1.2
duckduckgo_search==4.4.3
email-validator==2.1.0.post1
exchange_calendars==4.5.2
executing==2.0.1
fake-useragent==1.4.0
fastapi==0.109.2
feedparser==6.0.11
filelock==3.13.1
Flask==3.0.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.48.1
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.3
hyperframe==6.0.1
ibapi==9.81.1.post1
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langchain==0.1.5
langchain-community==0.0.18
langchain-core==0.1.19
langchain-openai==0.0.5
langsmith==0.0.86
llama-hub==0.0.78
llama-index==0.9.44
llama-parse==0.1.4
llmsherpa==0.1.4
lumibot==2.9.13
lumiwealth-tradier==0.1.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.3
multidict==5.2.0
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.11.1
openapi==1.1.0
packaging==23.2
pandas==2.0.0
pandas-datareader==0.10.0
pandas_market_calendars==4.4.0
parse==1.20.1
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
pdfkit==1.0.0
peewee==3.17.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.18.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.12.0
pyarrow==15.0.0
pycares==4.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pyee==8.2.2
Pygments==2.17.2
pyluach==2.2.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
Quandl==3.7.0
QuantStats==0.0.62
quantstats-lumi==0.1.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
retrying==1.3.4
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.2
scipy==1.10.1
seaborn==0.13.2
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.25
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.36.3
streamlit==1.31.0
streamlit-autorefresh==1.0.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
timedelta==2020.12.3
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.27.0.post1
validators==0.22.0
w3lib==2.1.2
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
# websockets==10.4
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yahoo-fin==0.8.9.1
yarl==1.9.4
yfinance==0.2.36
zipp==3.17.0
